+++ test/PerpetualIntelligence.Terminal.Test/Configuration/Options/TerminalOptionsTests.cs
@@ -23,6 +23,10 @@ namespace PerpetualIntelligence.Terminal.Configuration.Options
         {
             TerminalOptions options = new();
 
+            options.Id.Should().BeEmpty();
+            options.Name.Should().BeNull();
+            options.RootAsDriver.Should().BeFalse();
+
             options.Authentication.Should().NotBeNull();
             options.Checker.Should().NotBeNull();
             options.Extractor.Should().NotBeNull();